+++ apps/mobile/src/features/player/components/PlayerMainTab.tsx
@@ -3,6 +3,7 @@ import type { ImageRef } from 'expo-image'
 import { useRouter } from 'expo-router'
 import { memo, type RefObject } from 'react'
 import { StyleSheet, View } from 'react-native'
+import { ScrollView } from 'react-native-gesture-handler'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
 
 import useCurrentTrack from '@/hooks/player/useCurrentTrack'
@@ -31,7 +32,10 @@ const PlayerMainTab = memo(function PlayerMainTab({
 
 	if (!currentTrack) return null
 	return (
-		<View style={styles.container}>
+		<ScrollView
+			contentContainerStyle={styles.container}
+			showsVerticalScrollIndicator={false}
+		>
 			<TrackInfo
 				onArtistPress={() =>
 					currentTrack.artist?.remoteId
@@ -64,13 +68,13 @@ const PlayerMainTab = memo(function PlayerMainTab({
 					}}
 				/>
 			</View>
-		</View>
+		</ScrollView>
 	)
 })
 
 const styles = StyleSheet.create({
 	container: {
-		flex: 1,
+		flexGrow: 1,
 		justifyContent: 'space-between',
 	},
 	controlsContainer: {