+++ apps/mobile/src/features/playlist/remote/components/PlaylistItem.tsx
@@ -258,6 +258,7 @@ const styles = StyleSheet.create({
 		flexDirection: 'row',
 		alignItems: 'center',
 		marginTop: 2,
+		flexWrap: 'wrap',
 	},
 	dotSeparator: {
 		marginHorizontal: 4,